Build List with UITableViewController Intermediate Training Core Data
With our project now created and the skeleton of our app now slowly taking shape, we'll move to building a list to show some dummy data.  When it comes time to start saving objects into Core Data, we'll have a list ready to present everything inside our database.

If you've build a list using a UITableView before, you'll be quite familiar with the necessary steps of registering and dequeuing cells for our UITableViewController.  I've decided to use this component because it'll be simpler to perform edits and deletions later in the course.