Sending our Company Back through Delegation Intermediate Training Core Data
After a company is created, we need to send it back to our main list.  Doing this requires a connection to be made between our list and the creation form, this connection is what we call delegation.  In order to successfully know about the newly created object, we pass it back through a custom Protocol method call.  Upon success, we can update the UITableView by inserting a new row using an IndexPath.