Set Up Core Data and Save Company Intermediate Training Core Data
In today's lesson we'll first look at how to setup Core Data manually ourselves.  I like to do this because there aren't a lot of good tutorials online that show the correct steps for this process.  Once we have everything up and running with Core Data, we can create our Company entity in the editor.  Now there are some trickiness behind the auto generated files so do be careful.

Finally I'll go through how to initialize the Core Data stack so that we can save and fetch companies inside of our application.  This process isn't too difficult, just takes a few lines that might seem mysterious in the beginning.