Deleting Companies Intermediate Training Core Data
Now that we have our Company objects properly saved and presented in our application, it's time to allow the removal of these objects.  In this lesson, we will implement the ability to swipe on a cell to present various actions such as Edit and Delete.  Upon hitting the delete, we remove the object from our UITableView as well as deleting it from Core Data using our shared context.