Editing Company Entities Intermediate Training Core Data
Being able to edit objects is a big part of an application. Moreover, keeping your data consistent with all the changes happening throughout the lifecycle of your app is very important. In this lesson, we go ahead and implement an edit swipe action for our UITableView as well as edit the entity we've stored in Core Data.  Finally we'll need to refresh our main list if edits have been successfully saved.