Custom Company Cell Intermediate Training Core Data
Our UI currently looks a little strange with the company images and names aligned incorrectly.  Let's fix this issue by introducing our own Custom Company Cell  that will allow use to configure the cells a lot easier.  We'll be passing our company object to this class to minimize the code that is in our CompaniesController file.