Form Validation Intermediate Training Core Data
As with any type of application that takes user input, you'll need a way to verify that the inputs are actually correct.  One easy way to present a friendly message to your users is by using UIAlertController to present a small warning box if you've detected invalid form entries.  This process of checking user fields is typically referred to as Form Validation. Let's see this in action by introducing a birthday field for out Employees