UITableView Sections Filtering Intermediate Training Core Data
Many applications that you use on your iOS devices will contain lists that are grouped in certain sections. For example you can think of your address book where all your contacts are grouped under letters that their last names start with.  In today's lesson, we look at how UITableViews can be sectioned off for employees with short names and long names.

In doing so, we learn how we can apply a filter function on all employees to easily get the employees we want.  Finally, the lesson ends in providing a much more elegant solution with a two-dimensional array to support our sections.