UISegmentedControl Employee Types Intermediate Training Core Data
In this lesson, we'll introduce a grouping called Employee Type that help to break up our students into Executive, Senior Management, and Staff sessions.  To get started, let's add a UISegmentedControl in our Employee creation form.  Based on the selected segment, we'll save a property on Employees to help us perform the grouping later.