Employee Type Enumeration Grouping Intermediate Training Core Data
With the employee type now persisted in our Core Data database, we can now apply filtering based on employee.type.  Once we get this fix in, we'll be able to insert a new employee whenever we create one inside our application.

As a bonus, I'll talk about Enumeration and why it's useful to use them inside our programs.